More fixes..

+++ src/os_crypto/shared/sec.c
@@ -446,7 +446,7 @@ int CreateSecMSG(keystruct *keys, char *msg, char *msg_encrypted,
    
 
     snprintf(_tmpmsg, OS_MAXSTR,"%010u:%04hu:%05hu%0*d:%s", 
-                            curr_time, global_count, 
+                            (int)curr_time, global_count, 
                             rand1,  bfsize+1, 1, msg);
